use double precision only for wavelength
Wavelength needs to be in double precision in order to preserve
the precision in the measurements, but other quantities (esp.
flux) don't need double precision, and datamodel I/O has been
reverted to single-precision for these to save disk space.
This commit adapts to these changes.

+++ tests/test_SpectrumSet.py
@@ -21,12 +21,12 @@ class SpectrumSetTestCase(lsst.utils.tests.TestCase):
         # Spectrum
         self.length = 123
         self.fiberId = 456
-        self.image = self.rng.uniform(size=self.length).astype(float)
+        self.image = self.rng.uniform(size=self.length).astype(np.float32)
         self.mask = lsst.afw.image.Mask(self.length, 1)
         self.mask.array[:] = self.rng.randint(0, 2**30, self.length).astype(np.int32)
-        self.background = self.rng.uniform(size=self.length).astype(float)
-        self.norm = self.rng.uniform(size=self.length).astype(float)
-        self.covariance = self.rng.uniform(size=(3, self.length)).astype(float)
+        self.background = self.rng.uniform(size=self.length).astype(np.float32)
+        self.norm = self.rng.uniform(size=self.length).astype(np.float32)
+        self.covariance = self.rng.uniform(size=(3, self.length)).astype(np.float32)
         self.wavelengthArray = np.arange(1, self.length + 1, dtype=float)
 
     def makeSpectrum(self):
